ui: publish taxes and team settings pages
