+++ .changeset/taxes-team-settings-pages.md
@@ -1,0 +1,6 @@
+---
+"@voyantjs/finance-ui": minor
+"@voyantjs/admin": minor
+---
+
+Publish reusable TaxesPage and TeamSettingsPage settings compositions from their owning UI packages.
