+++ packages/admin/src/index.ts
@@ -47,6 +47,11 @@ export {
   type OperatorAdminUserMenuProps,
 } from "./components/operator-admin-user-menu.js"
 export {
+  TeamSettingsPage,
+  type TeamSettingsPageApi,
+  type TeamSettingsPageProps,
+} from "./components/team-settings-page.js"
+export {
   type AdminExtension,
   type AdminNavigationContribution,
   type AdminUiRouteContribution,
