+++ packages/finance-ui/src/i18n/en.ts
@@ -263,6 +263,121 @@ export const financeUiEn = {
       applied: "Applied",
     },
   },
+  taxesPage: {
+    title: "Tax Regimes and Classes",
+    description:
+      "Manage tax classes and the default tax regime each class uses for VAT calculation.",
+    addTax: "Add tax class",
+    empty: "No tax classes yet. Seed or create classes such as Normala, Redusa, and Art. 311.",
+    inactive: "Inactive",
+    edit: "Edit",
+    delete: "Delete",
+    deleteConfirm: "Delete this tax class?",
+    editSheetTitle: "Edit Tax Class",
+    newSheetTitle: "New Tax Class",
+    taxClassBadge: "Tax class",
+    defaultRegimeLabel: "Default regime",
+    regimeOverridesLabel: "Overrides",
+    regimeOverrideCount: "{count} overrides",
+    taxClassSectionTitle: "Tax class",
+    taxClassSectionDescription:
+      "The assignable classification stored on products and booking lines.",
+    defaultRegimeSectionTitle: "Default tax regime",
+    defaultRegimeSectionDescription:
+      "The VAT/tax calculation rule linked through tax_classes.default_regime_id.",
+    regimeOverridesSectionTitle: "Regime overrides",
+    regimeOverridesSectionDescription:
+      "Optional tax_classes.lines entries for base, addon, accommodation, or all lines.",
+    addRegimeOverride: "Add override",
+    removeRegimeOverride: "Remove override",
+    noRegimeOverrides: "No regime overrides. This class falls back to the default regime.",
+    appliesToLabel: "Applies to",
+    taxRegimeLabel: "Tax regime",
+    appliesToBase: "Base",
+    appliesToAddon: "Addon",
+    appliesToAccommodation: "Accommodation",
+    appliesToAll: "All",
+    taxClassLabelLabel: "Tax class label",
+    taxClassLabelPlaceholder: "Normala 21",
+    taxClassCodeLabel: "Tax class code",
+    taxClassCodePlaceholder: "normala",
+    taxClassDescriptionLabel: "Tax class description",
+    taxClassDescriptionPlaceholder: "Optional tax class description...",
+    regimeNameLabel: "Regime name",
+    regimeNamePlaceholder: "Normala",
+    regimeCodeLabel: "Regime code",
+    rateLabel: "Regime rate %",
+    jurisdictionLabel: "Jurisdiction",
+    legalReferenceLabel: "Legal reference",
+    legalReferencePlaceholder: "Optional legal reference",
+    regimeDescriptionLabel: "Regime description",
+    regimeDescriptionPlaceholder: "Optional tax regime description...",
+    activeLabel: "Active",
+    cancel: "Cancel",
+    saveChanges: "Save Changes",
+    createTax: "Create Tax Class",
+    validationNameRequired: "Tax class label is required",
+    validationRateInvalid: "Rate must be a non-negative number",
+    saveFailed: "Failed to save tax class",
+    policyTitle: "Tax Policy Profiles",
+    policyDescription:
+      "Manage implementation-specific tax rules that choose tax regimes automatically.",
+    addPolicyProfile: "Add profile",
+    addPolicyRule: "Add rule",
+    policyEmpty: "No tax policy profiles yet.",
+    policyRulesEmpty: "No policy rules yet. Add ordered rules for this profile.",
+    deletePolicyProfileConfirm: "Delete this tax policy profile and its rules?",
+    deletePolicyRuleConfirm: "Delete this tax policy rule?",
+    editPolicyProfileSheetTitle: "Edit Tax Policy Profile",
+    newPolicyProfileSheetTitle: "New Tax Policy Profile",
+    editPolicyRuleSheetTitle: "Edit Tax Policy Rule",
+    newPolicyRuleSheetTitle: "New Tax Policy Rule",
+    policyProfileNameLabel: "Profile name",
+    policyProfileNamePlaceholder: "Romanian travel operator",
+    policyProfileCodeLabel: "Profile code",
+    policyProfileCodePlaceholder: "ro-travel-operator",
+    policyProfileDescriptionLabel: "Profile description",
+    policyProfileDescriptionPlaceholder: "Optional policy profile description...",
+    policyPriorityLabel: "Priority",
+    policySideLabel: "Side",
+    policyRuleNameLabel: "Rule name",
+    policyRuleNamePlaceholder: "Accommodation in Romania",
+    policyConditionLabel: "Condition",
+    policyConditionSectionTitle: "Condition",
+    policyConditionSectionDescription:
+      "Choose when this rule applies. Rules are evaluated by priority.",
+    policyConditionModeLabel: "Match",
+    policyConditionAlways: "Always",
+    policyConditionAlwaysDescription:
+      "This rule always matches and is usually used as the last fallback rule.",
+    policyConditionModeAll: "All conditions",
+    policyConditionModeAny: "Any condition",
+    addPolicyCondition: "Add condition",
+    removePolicyCondition: "Remove condition",
+    policyFactLabel: "Fact",
+    policyFactHasAccommodation: "Has accommodation",
+    policyFactAccommodationCountries: "Accommodation countries",
+    policyOperatorLabel: "Operator",
+    policyOperatorEquals: "is",
+    policyOperatorContains: "contains",
+    policyValueLabel: "Value",
+    policyValueYes: "Yes",
+    policyValueNo: "No",
+    policyActionsLabel: "Actions",
+    policySideSell: "Sell",
+    policySideBuy: "Buy",
+    createPolicyProfile: "Create Profile",
+    createPolicyRule: "Create Rule",
+    validationPolicyProfileNameRequired: "Profile name is required",
+    validationPolicyProfileRequired: "Tax policy profile is required",
+    validationPolicyRuleNameRequired: "Rule name is required",
+    validationPolicyRuleRegimeRequired: "Tax regime is required",
+    validationPolicyRulePriorityInvalid: "Priority must be a non-negative integer",
+    validationPolicyRuleConditionInvalid:
+      "Condition must include at least one valid row. Country values must be two-letter ISO codes.",
+    savePolicyProfileFailed: "Failed to save tax policy profile",
+    savePolicyRuleFailed: "Failed to save tax policy rule",
+  },
   supplierPaymentDialog: {
     title: "Record Supplier Payment",
     fields: {
