+++ packages/finance-ui/src/i18n/messages.ts
@@ -41,6 +41,113 @@ export type SupplierPaymentMethod = (typeof supplierPaymentMethods)[number]
 export type SupplierPaymentStatus = SupplierPaymentRecord["status"]
 export type CreditNoteStatus = CreditNoteRecord["status"]
 
+export type TaxesPageMessageKey =
+  | "title"
+  | "description"
+  | "addTax"
+  | "empty"
+  | "inactive"
+  | "edit"
+  | "delete"
+  | "deleteConfirm"
+  | "editSheetTitle"
+  | "newSheetTitle"
+  | "taxClassBadge"
+  | "defaultRegimeLabel"
+  | "regimeOverridesLabel"
+  | "regimeOverrideCount"
+  | "taxClassSectionTitle"
+  | "taxClassSectionDescription"
+  | "defaultRegimeSectionTitle"
+  | "defaultRegimeSectionDescription"
+  | "regimeOverridesSectionTitle"
+  | "regimeOverridesSectionDescription"
+  | "addRegimeOverride"
+  | "removeRegimeOverride"
+  | "noRegimeOverrides"
+  | "appliesToLabel"
+  | "taxRegimeLabel"
+  | "appliesToBase"
+  | "appliesToAddon"
+  | "appliesToAccommodation"
+  | "appliesToAll"
+  | "taxClassLabelLabel"
+  | "taxClassLabelPlaceholder"
+  | "taxClassCodeLabel"
+  | "taxClassCodePlaceholder"
+  | "taxClassDescriptionLabel"
+  | "taxClassDescriptionPlaceholder"
+  | "regimeNameLabel"
+  | "regimeNamePlaceholder"
+  | "regimeCodeLabel"
+  | "rateLabel"
+  | "jurisdictionLabel"
+  | "legalReferenceLabel"
+  | "legalReferencePlaceholder"
+  | "regimeDescriptionLabel"
+  | "regimeDescriptionPlaceholder"
+  | "activeLabel"
+  | "cancel"
+  | "saveChanges"
+  | "createTax"
+  | "validationNameRequired"
+  | "validationRateInvalid"
+  | "saveFailed"
+  | "policyTitle"
+  | "policyDescription"
+  | "addPolicyProfile"
+  | "addPolicyRule"
+  | "policyEmpty"
+  | "policyRulesEmpty"
+  | "deletePolicyProfileConfirm"
+  | "deletePolicyRuleConfirm"
+  | "editPolicyProfileSheetTitle"
+  | "newPolicyProfileSheetTitle"
+  | "editPolicyRuleSheetTitle"
+  | "newPolicyRuleSheetTitle"
+  | "policyProfileNameLabel"
+  | "policyProfileNamePlaceholder"
+  | "policyProfileCodeLabel"
+  | "policyProfileCodePlaceholder"
+  | "policyProfileDescriptionLabel"
+  | "policyProfileDescriptionPlaceholder"
+  | "policyPriorityLabel"
+  | "policySideLabel"
+  | "policyRuleNameLabel"
+  | "policyRuleNamePlaceholder"
+  | "policyConditionLabel"
+  | "policyConditionSectionTitle"
+  | "policyConditionSectionDescription"
+  | "policyConditionModeLabel"
+  | "policyConditionAlways"
+  | "policyConditionAlwaysDescription"
+  | "policyConditionModeAll"
+  | "policyConditionModeAny"
+  | "addPolicyCondition"
+  | "removePolicyCondition"
+  | "policyFactLabel"
+  | "policyFactHasAccommodation"
+  | "policyFactAccommodationCountries"
+  | "policyOperatorLabel"
+  | "policyOperatorEquals"
+  | "policyOperatorContains"
+  | "policyValueLabel"
+  | "policyValueYes"
+  | "policyValueNo"
+  | "policyActionsLabel"
+  | "policySideSell"
+  | "policySideBuy"
+  | "createPolicyProfile"
+  | "createPolicyRule"
+  | "validationPolicyProfileNameRequired"
+  | "validationPolicyProfileRequired"
+  | "validationPolicyRuleNameRequired"
+  | "validationPolicyRuleRegimeRequired"
+  | "validationPolicyRulePriorityInvalid"
+  | "validationPolicyRuleConditionInvalid"
+  | "savePolicyProfileFailed"
+  | "savePolicyRuleFailed"
+
 export type FinanceUiMessages = {
   common: {
     cancel: string
@@ -271,6 +378,7 @@ export type FinanceUiMessages = {
     }
     creditNoteStatusLabels: Record<CreditNoteStatus, string>
   }
+  taxesPage: Record<TaxesPageMessageKey, string>
   supplierPaymentDialog: {
     title: string
     fields: {
