+++ packages/finance-ui/src/i18n/ro.ts
@@ -264,6 +264,122 @@ export const financeUiRo = {
       applied: "Aplicata",
     },
   },
+  taxesPage: {
+    title: "Regimuri si clase de taxe",
+    description:
+      "Administreaza clasele de taxe si regimul fiscal implicit folosit pentru calcul TVA.",
+    addTax: "Adauga clasa de taxe",
+    empty:
+      "Nu exista clase de taxe momentan. Seed sau creeaza clase precum Normala, Redusa si Art. 311.",
+    inactive: "Inactiv",
+    edit: "Editeaza",
+    delete: "Sterge",
+    deleteConfirm: "Stergi aceasta clasa de taxe?",
+    editSheetTitle: "Editeaza clasa de taxe",
+    newSheetTitle: "Clasa de taxe noua",
+    taxClassBadge: "Clasa de taxe",
+    defaultRegimeLabel: "Regim implicit",
+    regimeOverridesLabel: "Suprascrieri",
+    regimeOverrideCount: "{count} suprascrieri",
+    taxClassSectionTitle: "Clasa de taxe",
+    taxClassSectionDescription:
+      "Clasificarea selectabila salvata pe produse si linii de rezervare.",
+    defaultRegimeSectionTitle: "Regim fiscal implicit",
+    defaultRegimeSectionDescription:
+      "Regula de calcul TVA/taxe legata prin tax_classes.default_regime_id.",
+    regimeOverridesSectionTitle: "Suprascrieri de regim",
+    regimeOverridesSectionDescription:
+      "Intrari optionale tax_classes.lines pentru baza, addon, cazare sau toate liniile.",
+    addRegimeOverride: "Adauga suprascriere",
+    removeRegimeOverride: "Sterge suprascrierea",
+    noRegimeOverrides: "Nu exista suprascrieri de regim. Clasa foloseste regimul implicit.",
+    appliesToLabel: "Se aplica la",
+    taxRegimeLabel: "Regim fiscal",
+    appliesToBase: "Baza",
+    appliesToAddon: "Addon",
+    appliesToAccommodation: "Cazare",
+    appliesToAll: "Toate",
+    taxClassLabelLabel: "Eticheta clasei de taxe",
+    taxClassLabelPlaceholder: "Normala 21",
+    taxClassCodeLabel: "Cod clasa de taxe",
+    taxClassCodePlaceholder: "normala",
+    taxClassDescriptionLabel: "Descriere clasa de taxe",
+    taxClassDescriptionPlaceholder: "Descriere optionala pentru clasa de taxe...",
+    regimeNameLabel: "Nume regim",
+    regimeNamePlaceholder: "Normala",
+    regimeCodeLabel: "Cod regim",
+    rateLabel: "Cota regim %",
+    jurisdictionLabel: "Jurisdictie",
+    legalReferenceLabel: "Referinta legala",
+    legalReferencePlaceholder: "Referinta legala optionala",
+    regimeDescriptionLabel: "Descriere regim",
+    regimeDescriptionPlaceholder: "Descriere optionala pentru regimul fiscal...",
+    activeLabel: "Activ",
+    cancel: "Anuleaza",
+    saveChanges: "Salveaza modificarile",
+    createTax: "Creeaza clasa de taxe",
+    validationNameRequired: "Eticheta clasei de taxe este obligatorie",
+    validationRateInvalid: "Cota trebuie sa fie un numar pozitiv",
+    saveFailed: "Clasa de taxe nu a putut fi salvata",
+    policyTitle: "Profile de politica fiscala",
+    policyDescription:
+      "Administreaza reguli fiscale specifice implementarii care aleg automat regimurile fiscale.",
+    addPolicyProfile: "Adauga profil",
+    addPolicyRule: "Adauga regula",
+    policyEmpty: "Nu exista profile de politica fiscala momentan.",
+    policyRulesEmpty: "Nu exista reguli. Adauga reguli ordonate pentru acest profil.",
+    deletePolicyProfileConfirm: "Stergi acest profil de politica fiscala si regulile lui?",
+    deletePolicyRuleConfirm: "Stergi aceasta regula fiscala?",
+    editPolicyProfileSheetTitle: "Editeaza profilul de politica fiscala",
+    newPolicyProfileSheetTitle: "Profil de politica fiscala nou",
+    editPolicyRuleSheetTitle: "Editeaza regula fiscala",
+    newPolicyRuleSheetTitle: "Regula fiscala noua",
+    policyProfileNameLabel: "Nume profil",
+    policyProfileNamePlaceholder: "Operator turistic Romania",
+    policyProfileCodeLabel: "Cod profil",
+    policyProfileCodePlaceholder: "ro-travel-operator",
+    policyProfileDescriptionLabel: "Descriere profil",
+    policyProfileDescriptionPlaceholder: "Descriere optionala pentru profil...",
+    policyPriorityLabel: "Prioritate",
+    policySideLabel: "Parte",
+    policyRuleNameLabel: "Nume regula",
+    policyRuleNamePlaceholder: "Cazare in Romania",
+    policyConditionLabel: "Conditie",
+    policyConditionSectionTitle: "Conditie",
+    policyConditionSectionDescription:
+      "Alege cand se aplica aceasta regula. Regulile sunt evaluate dupa prioritate.",
+    policyConditionModeLabel: "Potrivire",
+    policyConditionAlways: "Intotdeauna",
+    policyConditionAlwaysDescription:
+      "Aceasta regula se potriveste intotdeauna si este folosita de obicei ca regula implicita finala.",
+    policyConditionModeAll: "Toate conditiile",
+    policyConditionModeAny: "Oricare conditie",
+    addPolicyCondition: "Adauga conditie",
+    removePolicyCondition: "Sterge conditia",
+    policyFactLabel: "Fapt",
+    policyFactHasAccommodation: "Are cazare",
+    policyFactAccommodationCountries: "Tari cazare",
+    policyOperatorLabel: "Operator",
+    policyOperatorEquals: "este",
+    policyOperatorContains: "contine",
+    policyValueLabel: "Valoare",
+    policyValueYes: "Da",
+    policyValueNo: "Nu",
+    policyActionsLabel: "Actiuni",
+    policySideSell: "Vanzare",
+    policySideBuy: "Achizitie",
+    createPolicyProfile: "Creeaza profilul",
+    createPolicyRule: "Creeaza regula",
+    validationPolicyProfileNameRequired: "Numele profilului este obligatoriu",
+    validationPolicyProfileRequired: "Profilul de politica fiscala este obligatoriu",
+    validationPolicyRuleNameRequired: "Numele regulii este obligatoriu",
+    validationPolicyRuleRegimeRequired: "Regimul fiscal este obligatoriu",
+    validationPolicyRulePriorityInvalid: "Prioritatea trebuie sa fie un intreg pozitiv",
+    validationPolicyRuleConditionInvalid:
+      "Conditia trebuie sa includa cel putin un rand valid. Tarile trebuie sa fie coduri ISO din doua litere.",
+    savePolicyProfileFailed: "Profilul de politica fiscala nu a putut fi salvat",
+    savePolicyRuleFailed: "Regula fiscala nu a putut fi salvata",
+  },
   supplierPaymentDialog: {
     title: "Inregistreaza Plata Furnizorului",
     fields: {
