+++ templates/operator/src/routes/_workspace/settings/taxes.tsx
@@ -1,6 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router"
-
-import { TaxesPage } from "@/components/voyant/settings/taxes-page"
+import { TaxesPage } from "@voyantjs/finance-ui"
 
 export const Route = createFileRoute("/_workspace/settings/taxes")({
   component: TaxesPage,
