+++ templates/operator/src/routes/_workspace/settings/team.tsx
@@ -1,6 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router"
-
-import { TeamSettingsPage } from "@/components/voyant/settings/team-settings-page"
+import { TeamSettingsPage } from "@voyantjs/admin"
 
 export const Route = createFileRoute("/_workspace/settings/team")({
   component: TeamSettingsPage,
